third_party/blink/web_tests/virtual/scoped-custom-element-registry-disabled/external/wpt/custom-elements/registries/CustomElementRegistry-initialize-expected.txt

This is a testharness.js-based test.
[FAIL] initialize is a function on both global and scoped CustomElementRegistry
  assert_equals: expected "function" but got "undefined"
[FAIL] initialize sets element.customElementRegistry to the global registry
  assert_equals: expected (object) null but got (undefined) undefined
[FAIL] initialize does not set the registry of nested shadow tree to the global registry
  assert_equals: expected (object) null but got (undefined) undefined
[FAIL] initialize sets element.customElementRegistry to a scoped registry
  assert_equals: expected (object) null but got (undefined) undefined
[FAIL] initialize does not set the registry of nested shadow tree to a scoped registry
  assert_equals: expected (object) null but got (undefined) undefined
[FAIL] initialize sets element.customElementRegistry permantently
  assert_equals: expected (object) null but got (undefined) undefined
[FAIL] initialize is no-op on a subtree with a non-null registry
  assert_equals: expected (object) object "[object CustomElementRegistry]" but got (undefined) undefined
[FAIL] initialize works on Document
  assert_equals: expected (object) null but got (undefined) undefined
[FAIL] initialize works on DocumentFragment
  assert_equals: expected (object) null but got (undefined) undefined
Harness: the test ran to completion.
